voice_function/requirements.txt
absl-py==2.2.2
actionlib==1.14.3
altgraph==0.17.4
angles==1.9.14
apturl==0.5.2
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
base-local-planner==1.17.3
bcrypt==3.1.7
bidict==0.23.1
blinker==1.8.2
bondpy==1.8.7
breezy==3.0.2
Brlapi==0.7.0
cachetools==5.5.2
camera-calibration==1.17.0
camera-calibration-parsers==1.12.1
catkin==0.8.12
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
cbor==1.0.0
certifi==2019.11.28
chardet==3.0.4
clarabel==0.9.0
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.3
coloredlogs==15.0.1
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.1
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cryptography==2.8
cupshelpers==1.0
cv-bridge==1.16.2
cvxpy==1.5.2
cycler==0.10.0
Cython==0.29.14
dash==3.0.4
dash-bootstrap-components==1.6.0
dash-daq==0.6.0
dash-iconify==0.1.2
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.7
diagnostic-analysis==1.12.1
diagnostic-common-diagnostics==1.12.1
diagnostic-updater==1.12.1
distro==1.4.0
distro-info==0.23+ubuntu1.1
dnspython==2.6.1
docutils==0.16
dulwich==0.19.15
duplicity==0.8.12.0
dynamic-reconfigure==1.7.6
ecos==2.0.14
empy==3.3.2
entrypoints==0.3
eventlet==0.39.1
Farama-Notifications==0.0.4
fasteners==0.14.1
fastimport==0.9.8
filelock==3.16.1
flask==3.0.3
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.0
future==0.18.2
gazebo-plugins==2.9.3
gazebo-ros==2.9.3
gencpp==0.7.2
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.1
gennodejs==2.0.2
genpy==0.6.18
google-auth==2.40.2
google-auth-oauthlib==1.0.0
gpg==1.13.1
greenlet==3.1.1
grpcio==1.70.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.0.0
h11==0.16.0
httplib2==0.14.0
humanfriendly==10.0
hyperlink==19.0.0
idna==2.8
image-geometry==1.16.2
importlib-metadata==8.5.0
importlib-resources==6.4.5
incremental==16.10.1
interactive-markers==1.12.2
itsdangerous==2.2.0
jinja2==3.1.6
joblib==1.4.2
joint-state-publisher==1.15.2
joint-state-publisher-gui==1.15.2
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
laser-geometry==1.6.8
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
message-filters==1.17.4
monotonic==1.5
mpi4py==3.0.3
mpmath==1.3.0
narwhals==1.39.0
nest-asyncio==1.6.0
netifaces==0.10.4
networkx==3.1
nose==1.3.7
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
olefile==0.46
onnx==1.17.0
onnxruntime==1.16.3
opencv-python==4.11.0.86
osqp==1.0.4
packaging==25.0
pandas==2.0.3
paramiko==2.6.0
pexpect==4.6.0
pillow==10.4.0
plotly==6.0.1
protobuf==5.29.4
psutil==5.5.1
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycryptodomex==3.6.1
pycups==1.9.73
pydot==1.4.1
pygame==2.6.1
PyGithub==1.43.7
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
PyQtWebEngine==5.14.0
pyRFC3339==1.1
PySide2==5.15.2.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-engineio==4.12.1
python-gitlab==2.0.1
python-gnupg==0.4.5
python-qt-binding==0.4.6
python-snappy==0.5.3
python-socketio==5.13.0
PyTrie==0.2
pytz==2025.2
pyxdg==0.26
PyYAML==6.0.2
qt-dotgraph==0.4.5
qt-gui==0.4.5
qt-gui-cpp==0.4.5
qt-gui-py-common==0.4.5
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==2.0.0
requests-unixsocket==0.2.0
resource-retriever==1.12.10
retrying==1.3.4
roman==2.0.0
rosbag==1.17.4
rosboost-cfg==1.15.10
rosclean==1.15.10
roscreate==1.15.10
rosdep==0.25.1
rosdep-modules==0.25.1
rosdistro==1.0.1
rosdistro-modules==1.0.1
rosgraph==1.17.4
rosinstall==0.7.8
rosinstall-generator==0.1.23
roslaunch==1.17.4
roslib==1.15.10
roslint==0.12.0
roslz4==1.17.4
rosmake==1.15.10
rosmaster==1.17.4
rosmsg==1.17.4
rosnode==1.17.4
rosparam==1.17.4
rospkg==1.6.0
rospkg-modules==1.6.0
rospy==1.17.4
rospy-message-converter==0.5.9
rosservice==1.17.4
rostest==1.17.4
rostopic==1.17.4
rosunit==1.15.10
roswtf==1.17.4
rqt-action==0.4.11
rqt-bag==0.5.3
rqt-bag-plugins==0.5.3
rqt-console==0.4.14
rqt-dep==0.4.14
rqt-graph==0.4.16
rqt-gui==0.5.5
rqt-gui-py==0.5.5
rqt-image-view==0.4.19
rqt-launch==0.4.10
rqt-logger-level==0.4.13
rqt-moveit==0.5.13
rqt-msg==0.4.12
rqt-nav-view==0.5.8
rqt-plot==0.4.16
rqt-pose-view==0.5.13
rqt-publisher==0.4.12
rqt-py-common==0.5.5
rqt-py-console==0.4.12
rqt-reconfigure==0.5.7
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-robot-steering==0.5.14
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.2
rqt-service-caller==0.4.12
rqt-shell==0.4.13
rqt-srv==0.4.11
rqt-tf-tree==0.6.5
rqt-top==0.4.11
rqt-topic==0.4.15
rqt-web==0.4.11
rsa==4.9.1
rviz==1.14.26
scikit-learn==1.3.2
scipy==1.10.1
screen-resolution-extra==0.0.0
scs==3.2.7.post2
SecretStorage==2.3.1
sensor-msgs==1.13.2
service-identity==18.1.0
shiboken2==5.15.2.1
Shimmy==2.0.0
simple-websocket==1.1.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smach==2.5.3
smach-ros==2.5.3
smclib==1.8.7
stable-baselines3==2.4.1
sympy==1.13.3
systemd-python==234
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tf==1.13.4
tf-conversions==1.13.4
tf2-geometry-msgs==0.7.10
tf2-kdl==0.7.10
tf2-py==0.7.10
tf2-ros==0.7.10
threadpoolctl==3.5.0
topic-tools==1.17.4
torch==2.4.1
transforms3d==0.4.2
triton==3.0.0
Twisted==18.9.0
txaio==2.10.0
typing-extensions==4.13.2
tzdata==2025.2
u-msgpack-python==2.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vcstools==0.1.42
wadllib==1.3.3
websocket-client==0.53.0
werkzeug==3.0.6
wrapt==1.11.2
wsaccel==0.6.2
wsproto==1.2.0
wstool==0.1.18
xacro==1.14.20
xkit==0.0.0
zipp==3.20.2
zope.interface==4.7.1
